Create Merge_tech_replicates.py
This script merges the technical replicates from the the log2 transformed data. The technical replicate cells are first defined for each sample and group. Then a new dataframe is created to store the averaged values. A value is averaged by the number of technical replicates present. If there is only 1 value for a protein group, this is taken forward into the new data frame without averaging. The first few rows of the new data frame are displayed and the file path is saved.