Create P_tests_FDH_corrected.py
After log2 transformation and defining the relevant groups, this script filters rows by a data completeness cut off (e.g., in this case at least 8 values per comparison per group). It then calculates the mean per group and the difference between the two. 

The states models implements multiple unpaired t-tests on the dataset that are then corrected using the FDH method with alpha 0.05. 

The output file dataframe lists the metadata for each row, the mean values, the differences, uncorrected and corrected P values, and a column stating whether the difference was significant in binary (yes or no). 

The script can be adjusted and modified as desired.